+++ messages/pt-BR/models_FileSystemItem.php
@@ -1,8 +1,7 @@
 <?php
-
-return [
-    'Downloads' => 'Transferências',
-    'Is Public' => 'É Público',
-    'Note: Changes of the folders visibility, will be inherited by all contained files and folders.' => 'Nota: Mudanças na visibilidade das pastas serão herdadas por todos os arquivos e pastas contidas nelas.',
-    'Hide in Stream' => '',
-];
+return array (
+  'Downloads' => 'Transferências',
+  'Hide in Stream' => 'Ocultar no Stream',
+  'Is Public' => 'É público',
+  'Note: Changes of the folders visibility, will be inherited by all contained files and folders.' => 'Observação: as alterações na visibilidade das pastas serão herdadas por todos os arquivos e pastas contidos.',
+);